fix/ refactor and bug fixing
