+++ src/components/CustomButton.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Button as MantineButton } from '@mantine/core';
-import { CustomButtonProps } from '../../types/types'; // Import the interface
+import { CustomButtonProps } from '../types/types'; // Import the interface
 
 const CustomButton: React.FC<CustomButtonProps> = ({ children, svgIcon, ...props }) => {
   return (
