+++ src/pages/_app.tsx
@@ -1,7 +1,7 @@
 import '../styles/globals.css';
 import '@mantine/core/styles.css';
 import { createTheme, MantineProvider } from '@mantine/core';
-import TodoApp from '@/pages/todo/TodoApp';
+import type { AppProps } from "next/app";
 
 if (process.env.NEXT_PUBLIC_API_MOCKING === 'true') {
   import('../mocks').then(({ setupMocks }) => {
@@ -10,31 +10,27 @@ if (process.env.NEXT_PUBLIC_API_MOCKING === 'true') {
 }
 
 const theme = createTheme({
-  fontFamily: 'Open Sans, sans-serif',
-  primaryColor: 'cyan',
   colors: {
     customPurple: [
-      '#f3e8ff', 
+      '#f3e8ff',
       '#e2ccff',
       '#d1b0ff',
       '#bf94ff',
       '#ae78ff',
-      '#9d5dff', 
+      '#9d5dff',
       '#723fff',
       '#6034cc',
       '#4e29a3',
-      '#3c1f7a'  
+      '#3c1f7a'
     ],
   },
 });
 
 
-const App = () => {
-  return (
+export default function App({ Component, pageProps }: AppProps) {
+  return <>
     <MantineProvider theme={theme}>
-      <TodoApp />
+      <Component {...pageProps} />
     </MantineProvider>
-  );
-};
-
-export default App;
+  </>
+}
